+++ source/utility.js
@@ -12,17 +12,18 @@ export function getPassThroughProps(props, excludeProps) {
 	return rest
 }
 
-export function scrollTo(node) {
+export function scrollTo(node, options) {
 	// https://github.com/stipsan/scroll-into-view-if-needed/issues/359
 	// scrollIntoView(ReactDOM.findDOMNode(this.field.current), false, { duration: 300 })
 	// Using `block: "center"` instead of `block: "nearest"`
 	// because otherwise the `node` might be obstructed by a "floating" header.
 	// https://github.com/stipsan/scroll-into-view-if-needed/issues/126#issuecomment-533089870
 	scrollIntoView(node, {
-		scrollMode : 'if-needed',
-		behavior   : 'smooth',
-		block      : 'center',
-		inline     : 'nearest'
+		scrollMode: 'if-needed',
+		behavior: 'smooth',
+		block: 'center',
+		inline: 'nearest',
+		duration: options && options.duration
 	})
 }